Fix prestissimo to ignore curl error message unless error code is set
There's currently a bug in curl 7.64 where it puts host not resolved
errors in the error buffer even though the host has been found and
the download was successful.

This bug is causing prestissimo to claim that it failed to download
most of its packages, thus causing composer to download them again
serially.

While the bug is in curl, according to
https://curl.haxx.se/libcurl/c/CURLOPT_ERRORBUFFER.html, the error
buffer should be ignored unless an error code is returned, so
prestissimo should be fixed to do so.  This patch contains the fix.

Signed-off-by: Jonathan Dieter <jonathan.dieter@spearline.com>